# public/openapi.yaml
openapi: 3.1.0
info:
  title: Lune API
  version: 0.3.0
  description: >
    API publique (auth, santé) et API app (PERSONAL / PRO). Toutes les IDs BigInt sont sérialisées en string.
    Toutes les réponses mutables exigent CSRF (Origin) et renvoient `x-request-id` + `cache-control: no-store`.
servers:
  - url: http://localhost:3000
    description: Développement local
  - url: https://diwanbg.work
    description: Production (Railway)
security:
  - cookieAuth: []
...
paths:
...
  /api/pro/businesses/{businessId}/projects/{projectId}:
    get:
      tags: [Pro / Projects]
      security: [ { cookieAuth: [] } ]
      summary: Détail projet
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Projet
          content:
            application/json:
              schema:
                type: object
                properties: { item: { $ref: '#/components/schemas/Project' } }
    patch:
      tags: [Pro / Projects]
      security: [ { cookieAuth: [] } ]
      summary: Met à jour un projet
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdate' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    delete:
      tags: [Pro / Projects]
      security: [ { cookieAuth: [] } ]
      summary: Supprime un projet
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    BusinessId:
      name: businessId
      in: path
      required: true
      schema: { type: string }
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Ok:
      description: Opération réussie
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: true }
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        businessId: { type: string }
        clientId: { type: string, nullable: true }
        categoryReferenceId: { type: string, nullable: true }
        categoryReferenceName: { type: string, nullable: true }
        tagReferences:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
        name: { type: string }
        status: { type: string }
        quoteStatus: { type: string }
        depositStatus: { type: string }
        createdAt: { type: string, format: date-time }
    ProjectUpdate:
      type: object
      properties:
        name: { type: string }
        status: { type: string }
        categoryReferenceId: { type: string, nullable: true }
        tagReferenceIds:
          type: array
          items: { type: string }
        startDate: { type: string, format: date-time, nullable: true }
        endDate: { type: string, format: date-time, nullable: true }
...